Create StudentRegistration.py
This program is to get student details from the user of the program and then display information on the screen.

Display a main heading
Obtain information about each student including student ID, name, the course in which they are enrolled, the course fee.
display their details on the screen.
After inputting the details for the three students , display their details with heading, and add all the course fee as total fee.